Added logging, added unittesting

+++ webapp/src/app.py
@@ -2,6 +2,13 @@ import os
 import sys
 import json
 import logging
+
+logging.basicConfig(
+    format="%(asctime)s : %(filename)s : %(threadName)s_%(thread)d(%(funcName)s)[LINE:%(lineno)d] : %(levelname)s : %(message)s",
+    level=logging.DEBUG,
+    filename='/var/log/webapp/flask.log',
+    filemode='w')
+
 from flask import Flask, flash, render_template, request, redirect, jsonify, Response
 from forms import SearchForm
 
@@ -18,24 +25,23 @@ app.secret_key = "flask rocks!"
 
 @app.route('/', methods = ['POST', 'GET'])
 def index():
-    search = SearchForm(request.form)
+    search = SearchForm()
     if request.method == 'POST':
-        return search_result(search)
+        return search_result()
     return render_template('index.html', form=search)
 
-@app.route('/results', methods = ['POST', 'GET'])
-def search_result(search):
+def search_result():
     form = SearchForm()
-    search_string = form.search.data
-
-    results = []
-    results = DATABASE.search_text(search_string)
+    results = DATABASE.search_text(form.search.data)
+    logging.info("WEB Search: data - %s", form.search.data)
 
     if not results:
-        flash('No results found!')
-        return redirect('/')
+        not_found = [('null', 'Not found', 'null', 'null')]
+        logging.info("WEB Search: No data was found in DB")
+        return render_template('index.html', form=form, results=not_found)
     else:
-        return render_template('index.html', form=search, results=results)
+        logging.info("WEB Search: Found - %s datas in DB", len(results))
+        return render_template('index.html', form=form, results=results)
 
 @app.route('/api_search', methods = ['POST', 'GET'])
 def api_search():
@@ -43,25 +49,43 @@ def api_search():
         Посылает json ответ
     """
     data = request.args.get('data')
-    results = DATABASE.search_text(data)
+    if data:
+        logging.info("API Search: Data - %s", data)
+        results = DATABASE.search_text(data)
 
-    if results:
-        main_json = []
+        if results:
+            logging.info("API Search: Found - %s datas in DB", len(results))
+            main_json = []
 
-        for res in results:
-            datetime = res[2].strftime("%Y-%m-%d %H:%M:%S")
-            temp = {
-                "id" : res[0],
-                "text" : res[1],
-                "created_date" : datetime,
-                "rubrics" : res[3]
-            }
-            main_json.append(temp)
-        return Response(json.dumps(main_json, ensure_ascii=False).encode('utf8'), mimetype='application/json')
+            for res in results:
+                datetime = res[2].strftime("%Y-%m-%d %H:%M:%S")
+                temp = {
+                    "id" : res[0],
+                    "text" : res[1],
+                    "created_date" : datetime,
+                    "rubrics" : res[3]
+                }
+                main_json.append(temp)
+            return Response(
+                json.dumps(main_json, ensure_ascii=False).encode('utf8'),
+                mimetype='application/json'
+            )
+
+        else:
+            logging.info("API Search: Not Found any data in DB")
+            err_json = {"error" : "not_found"}
+            return Response(
+                json.dumps(err_json, ensure_ascii=False).encode('utf8'),
+                mimetype='application/json'
+            )
 
     else:
-        err_json = {"error" : "not_found"}
-        return Response(json.dumps(err_json, ensure_ascii=False).encode('utf8'), mimetype='application/json')
+        logging.info("API Search: Data wasn't sent")
+        err_json = {"error" : "give_me_data"}
+        return Response(
+            json.dumps(err_json, ensure_ascii=False).encode('utf8'),
+            mimetype='application/json'
+        )
 
 
 @app.route('/del', methods = ['POST', 'GET'])
@@ -70,14 +94,27 @@ def delete_item():
         http://localhost:8888/del?id=38
     """
     data = request.args.get('id')
-    result = DATABASE.delete_row(data)
 
-    if reusult == 0:
-        answer = 'deleted'
+    if data:
+        logging.info("API Deletion: id - %s", data)
+        result = DATABASE.delete_row(data)
     else:
-        anwer = 'deletion_error'
+        logging.info("API Deletion: No ID was sent")
+        result = 2
 
-    return Response(json.dumps({"result": answer}, ensure_ascii=False).encode('utf8'), mimetype='application/json')
+    if result == 0:
+        logging.info("API Deletion: id - %s. Successfull deleted", data)
+        answer = 'deleted'
+    elif result == 2:
+        answer = 'give_me_id'
+    else:
+        logging.info("API Deletion: ID - %s wasn't found in DB", data)
+        answer = 'id_not_found'
+
+    return Response(
+        json.dumps({"result": answer}, ensure_ascii=False).encode('utf8'),
+        mimetype='application/json'
+        )
 
 
 if __name__ == '__main__':
